Reorganize the code under analyzer/.
This moves some disparate parts out of scan.go into {compare,graph}.go.
